0.17.5 drafted - implemented storage querying
