+++ Cargo.toml
@@ -7,7 +7,7 @@ repository = "https://github.com/insight-platform/Similari"
 readme = "README.md"
 keywords = ["machine-learning", "similarity", "tracking"]
 categories = ["algorithms", "data-structures", "database", "mathematics", "science"]
-version = "0.17.4"
+version = "0.17.5"
 edition = "2021"
 license="Apache-2.0"
 rust-version = "1.62"
